requirements.txt
matplotlib==3.4.3
numpy==1.21.2
opencv_python_headless==4.5.4.58
pandas==1.3.4
Pillow==9.0.0
PyInputPlus==0.2.12
tqdm==4.61.2
